redundant casts

+++ sandbox/PerfBenchmark/TargetClass.cs
@@ -59,12 +59,12 @@ namespace PerfBenchmark
                 {
                     Number1 = (sbyte)random.Next(),
                     Number2 = (short)random.Next(),
-                    Number3 = (int)random.Next(),
+                    Number3 = random.Next(),
                     Number4 = (long)new LongUnion { Int1 = random.Next(), Int2 = random.Next() }.Long,
                     Number5 = (byte)random.Next(),
                     Number6 = (ushort)random.Next(),
                     Number7 = (uint)random.Next(),
-                    Number8 = (ulong)new LongUnion { Int1 = random.Next(), Int2 = random.Next() }.Long,
+                    Number8 = new LongUnion { Int1 = random.Next(), Int2 = random.Next() }.Long,
                 };
             }
         }
