+++ src/Utf8Json/JsonWriter.cs
@@ -294,7 +294,7 @@ namespace Utf8Json
 #endif
         public void WriteByte(byte value)
         {
-            WriteUInt64((ulong)value);
+            WriteUInt64(value);
         }
 
 #if NETSTANDARD
@@ -302,7 +302,7 @@ namespace Utf8Json
 #endif
         public void WriteUInt16(ushort value)
         {
-            WriteUInt64((ulong)value);
+            WriteUInt64(value);
         }
 
 #if NETSTANDARD
@@ -310,7 +310,7 @@ namespace Utf8Json
 #endif
         public void WriteUInt32(uint value)
         {
-            WriteUInt64((ulong)value);
+            WriteUInt64(value);
         }
 
         public void WriteUInt64(ulong value)
@@ -323,7 +323,7 @@ namespace Utf8Json
 #endif
         public void WriteSByte(sbyte value)
         {
-            WriteInt64((long)value);
+            WriteInt64(value);
         }
 
 #if NETSTANDARD
@@ -331,7 +331,7 @@ namespace Utf8Json
 #endif
         public void WriteInt16(short value)
         {
-            WriteInt64((long)value);
+            WriteInt64(value);
         }
 
 #if NETSTANDARD
@@ -339,7 +339,7 @@ namespace Utf8Json
 #endif
         public void WriteInt32(int value)
         {
-            WriteInt64((long)value);
+            WriteInt64(value);
         }
 
         public void WriteInt64(long value)
